Fix active unit check during StaffAI_TargetArea (#711)
Active unit id is not correctly loaded when finding a valid tile for an enemy staff user with a relevant staff (e.g. Fortify). This can cause the routine to identify the unit's current tile as being occupied by another unit, which then makes the routine check the next tile. If the user has 0 Move, Fortify won't be used because no valid tile will be found.